Enabled forgery protection for test environment
Because: The CRSF token is used by javascript, which causes a javascript
error in the test environment and fails any tests that rely on
javascript code to pass.

This commit: Enables forgery protection in the test environment.

js_errors have also been enabled in cuprite so any future js errors are
re-reraised in ruby